requirements.txt
# Django
Django==3.2.9
django-allauth~=0.43.0

# Envs
django-dotenv==1.4.2

# DRF
djangorestframework==3.12.4
django-filter==2.4.0
django-rest-auth==0.9.5

# Celery
celery==5.2.1

# Database
psycopg2-binary==2.9.2
dj_database_url==0.5.0

#Server
gunicorn==20.0.4
